// FRONTEND/chatbot-admin-panel/src/services/api/schemas/adminPanelMenuResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { AdminPanelMenuResponseIconName } from './adminPanelMenuResponseIconName';
import type { AdminPanelMenuResponseParentId } from './adminPanelMenuResponseParentId';

export interface AdminPanelMenuResponse {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 255
   */
  frontend_route: string;
  icon_name?: AdminPanelMenuResponseIconName;
  parent_id?: AdminPanelMenuResponseParentId;
  display_order?: number;
  id: number;
  created_at: string;
  updated_at: string;
}