// FRONTEND/chatbot-admin-panel/src/services/api/schemas/adminPanelMenuCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { AdminPanelMenuCreateIconName } from './adminPanelMenuCreateIconName';
import type { AdminPanelMenuCreateParentId } from './adminPanelMenuCreateParentId';

export interface AdminPanelMenuCreate {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 255
   */
  frontend_route: string;
  icon_name?: AdminPanelMenuCreateIconName;
  parent_id?: AdminPanelMenuCreateParentId;
  display_order?: number;
}