// FRONTEND/chatbot-admin-panel/src/services/api/schemas/adminPanelMenuUpdate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { AdminPanelMenuUpdateName } from './adminPanelMenuUpdateName';
import type { AdminPanelMenuUpdateFrontendRoute } from './adminPanelMenuUpdateFrontendRoute';
import type { AdminPanelMenuUpdateIconName } from './adminPanelMenuUpdateIconName';
import type { AdminPanelMenuUpdateParentId } from './adminPanelMenuUpdateParentId';
import type { AdminPanelMenuUpdateDisplayOrder } from './adminPanelMenuUpdateDisplayOrder';

export interface AdminPanelMenuUpdate {
  name?: AdminPanelMenuUpdateName;
  frontend_route?: AdminPanelMenuUpdateFrontendRoute;
  icon_name?: AdminPanelMenuUpdateIconName;
  parent_id?: AdminPanelMenuUpdateParentId;
  display_order?: AdminPanelMenuUpdateDisplayOrder;
}